shooter score assignment

+++ SemperPrecisStageTracker.Models/Aggregation/ShooterMatchResult.cs
@@ -12,6 +12,6 @@ namespace SemperPrecisStageTracker
     public class ShooterStageResult
     {
         public int StageIndex { get; set; }
-        public int Total { get; set; }
+        public decimal Total { get; set; }
     }
 }
